+++ cv_Point.h
@@ -5,7 +5,7 @@ public:
 	int x;
 	int y;
 
-	double DistanceCalculation(cv_Point dst) const;
+	float DistanceCalculation(cv_Point dst) const;
 	cv_Point();
 	cv_Point(int _x, int _y);
 	~cv_Point();
